Helpers for resolving nuget packages
