+++ src/DotNetProjectFile.Analyzers/MsBuild/PackageReference.cs
@@ -8,45 +8,4 @@ public sealed class PackageReference(XElement element, Node parent, MsBuildProje
     public string VersionOrVersionOverride => Version ?? VersionOverride ?? string.Empty;
 
     public string? PrivateAssets => Attribute() ?? Child();
-
-    public (Node Node, string Version)? ResolveVersionVerbose()
-    {
-        if (Project.ManagePackageVersionsCentrally() is true)
-        {
-            if (VersionOverride is { Length: > 0 })
-            {
-                return (this, VersionOverride);
-            }
-
-            var versionOverrideNode = Project
-                .WalkBackward()
-                .OfType<PackageReference>()
-                .FirstOrDefault(p => p.IncludeOrUpdate == IncludeOrUpdate);
-
-            if (versionOverrideNode?.VersionOverride is { Length: > 0 } versionOverride)
-            {
-                return (versionOverrideNode, versionOverride);
-            }
-
-            var versionNode = Project
-                .WalkBackward()
-                .OfType<PackageVersion>()
-                .FirstOrDefault(v => v.Include == IncludeOrUpdate);
-
-            if (versionNode?.Version is { Length: > 0 } version)
-            {
-                return (versionNode, version);
-            }
-        }
-        else if (Version is { Length: > 0 })
-        {
-            return (this, Version);
-        }
-
-        return null;
-    }
-
-    /// <summary>Resolves the version taking CPM into account.</summary>
-    public string? ResolveVersion()
-        => ResolveVersionVerbose()?.Version;
 }
